Include defaultFonts.ts in sideEffects
Otherwise ESBuild would just drop its imports in the bundle.
The purpose of importing defaultFonts is to include the
.woff/.woff2 files of the font in the output, which is side-effecting so
it should be marked as such.

This was working anyway because defaultFonts is also its own entrypoint
so the files were included regardless.
That said, if in the future we were to remove the defaultFonts entrypoint,
it would've resulted in the files not showing up in the output anymore.